Create common script for Atrium and Artery problem
